+++ AssemblyInfo.cs
@@ -28,11 +28,11 @@ using AT_Utils;
 // and "{Major}.{Minor}.{Build}.*" will update just the revision.
 
 #if NIGHTBUILD
-[assembly: AssemblyVersion("3.7.*")]
+[assembly: AssemblyVersion("3.8.*")]
 #else
-[assembly: AssemblyVersion("3.7.2")]
+[assembly: AssemblyVersion("3.8.0")]
 #endif
-[assembly: KSPAssembly("ThrottleControlledAvionics", 3, 7)]
+[assembly: KSPAssembly("ThrottleControlledAvionics", 3, 8)]
 
 // The following attributes are used to specify the signing key for the assembly, 
 // if desired. See the Mono documentation for more information about signing.
@@ -46,8 +46,8 @@ namespace ThrottleControlledAvionics
     {
         public ModInfo()
         {
-            MinKSPVersion = new Version(1, 9, 0);
-            MaxKSPVersion = new Version(1, 10, 0);
+            MinKSPVersion = new Version(1, 11, 1);
+            MaxKSPVersion = new Version(1, 11, 1);
 
             VersionURL = "https://raw.githubusercontent.com/allista/ThrottleControlledAvionics/master/GameData/ThrottleControlledAvionics/ThrottleControlledAvionics.version";
             UpgradeURL = "http://spacedock.info/mod/198/Throttle%20Controlled%20Avionics";