+++ VesselProps/EnginesProps.cs
@@ -58,6 +58,8 @@ namespace ThrottleControlledAvionics
         public float    MaxMassFlow { get; private set; }
         public float    MaxVe { get; private set; } //Specific impulse in m/s, aka effective exhaust velocity
 
+        public float WeightedThrustMod { get; private set; }
+
         public override void Clear() { All.Clear(); RCS.Clear(); }
 
         public bool AddEngine(PartModule pm) 
@@ -117,7 +119,7 @@ namespace ThrottleControlledAvionics
             for(int i = 0; i < NumActive; i++) 
             {
                 var e = Active[i];
-                if(e.isVSC)
+                if(e.isThruster)
                 {
                     float mFlow;
                     thrust += e.ThrustAtAlt(vel, alt, out mFlow); 
@@ -463,19 +465,24 @@ namespace ThrottleControlledAvionics
             MaxDefThrust = Vector3.zero;
             MaxThrust = Vector3.zero;
             MaxMassFlow = 0f;
+            WeightedThrustMod = 0f;
             Slow = false;
             var total_thrust = 0f;
+            var total_controllable_thrust = 0f;
             for(int i = 0; i < NumActive; i++) 
             {
                 var e = Active[i];
-                if(e.isVSC)
+                if(e.isThruster)
                 {
                     var thrust = e.nominalCurrentThrust(e.limit);
                     total_thrust += thrust;
+                    if(!e.thrustLimiterLocked)
+                        total_controllable_thrust += thrust;
                     MaxDefThrust += e.defThrustDir*thrust;
                     MaxThrust += e.wThrustDir*thrust;
                     MaxMassFlow += e.MaxFuelFlow*e.limit;
-                    if(e.useEngineResponseTime && e.finalThrust > 0)
+                    WeightedThrustMod += e.thrustMod * thrust;
+                    if(!e.thrustLimiterLocked && e.useEngineResponseTime && e.finalThrust > 0)
                     {
                         if(e.engineDecelerationSpeed > 0)
                             DecelerationSpeed += thrust*e.engineDecelerationSpeed;
@@ -487,17 +494,20 @@ namespace ThrottleControlledAvionics
             }
             if(AccelerationSpeed > 0)
             { 
-                AccelerationSpeed /= total_thrust; 
+                AccelerationSpeed /= total_controllable_thrust; 
                 AccelerationTime90 = Utils.LerpTime(AccelerationSpeed, 0.9f);
                 Slow = true;
             }
             if(DecelerationSpeed > 0) 
             { 
-                DecelerationSpeed /= total_thrust; 
+                DecelerationSpeed /= total_controllable_thrust; 
                 DecelerationTime10 = Utils.LerpTime(DecelerationSpeed, 0.9f);
                 Slow = true;
             }
-            if(MassFlow > MaxMassFlow) MaxMassFlow = MassFlow;
+            if(MassFlow > MaxMassFlow)
+                MaxMassFlow = MassFlow;
+            if(WeightedThrustMod > 0)
+                WeightedThrustMod /= total_thrust;
             MaxThrustM = MaxThrust.magnitude;
             MaxVe = MaxThrustM/MaxMassFlow;
             MaxAccel = MaxThrustM/VSL.Physics.M;