+++ make-nightbuild.sh
@@ -2,7 +2,7 @@
 
 cd $(dirname "$0")
 
-../../PyKSPutils/make_mod_release \
+make_mod_release \
 -e '*/config.xml' '*/TCA.conf' '*/TCA.macro' '*.user' '*.orig' \
 '*/001_AnisotropicPartResizer*' \
 '*/002_MultiAnimators.dll' \